+++ packages/winnow/src/query.js
@@ -23,7 +23,7 @@ function params (features, options) {
   const params = []
   // NOTE: order matters here
   // Fields stage
-  if (options.projection) params.push(options.projection)
+  if (options.projection && !options.aggregates) params.push(options.projection)
   if (options.geometryPrecision) params.push(options.geometryPrecision)
   // From stage
   params.push(Array.isArray(features) ? features : [features])