+++ packages/winnow/CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+### Unreleased
+### Fixed
+* Query ignores projection parameter when getting an aggregate
+
 ## [1.10.3] - 06-26-2017
 ### Fixed
 Deep clone for dates since mutating dates has down stream consequences if reused